backend/tests/testing_lang/requirements.txt
langchain==0.3.27
langchain-openai==0.3.33
langgraph==0.6.7
langchain-tavily==0.2.11
pydantic==2.11.9
tiktoken==0.11.0
python-dotenv==1.1.1
fastapi==0.117.1
uvicorn==0.37.0
loguru==0.7.3